scan: only accept tags in the first column
There is an unnecessary line = line.strip() which causes Reviewed-by:
and Acked-by: tags to be accepted from quoted email messages or commits.
This leads to spurious tags.

Since parse_tag() already strips the value, there is no need to strip
the line.  Valid tags must begin in the first column.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>