Python 3 conversion
This patch converts the code to Python 3, which is now the default on
new operating systems since Python 2 will not be maintained past 2020.

The client portion of the code has been tested through everyday use over
the past few months.

The server portion of the code has not been tested yet and may require
additional fixes.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>